Pull ShaderLibraryIndex out to its own file.
