+++ src/commonMain/kotlin/baaahs/libraries/ShaderLibrary.kt
@@ -27,22 +27,3 @@ data class ShaderLibrary(
         }
     }
 }
-
-/** On-disk representation of [ShaderLibrary]. */
-@Serializable
-data class ShaderLibraryIndex(
-    val title: String,
-    val description: String?,
-    val license: String?,
-    val entries: List<Entry> = emptyList()
-) {
-    @Serializable
-    data class Entry(
-        val id: String,
-        val title: String,
-        val description: String?,
-        val lastModifiedMs: Time,
-        val tags: List<String>,
-        val srcFile: String
-    )
-}
